Fix fragment delimiters strategy
It was a dumb idea to try to rely on keeping a cache entry for every Fragment,
since their positions in the text can obviously change when another Fragment
gets inserted before. Caching the whole list of delimiters and invalidating it when
there is an update is a much saner strategy.